elements/formFields.js: moved from lib/
